FIX remove bindless logic - problematic with intel Arc - won't port well to vulkan - too much trouble maintaining 2 implementations in shaders - significantly smaller material buffer size

+++ src/material/ArrayTexture.h
@@ -34,7 +34,8 @@ public:
         ArrayTexture& arr,
         uint32_t layer) override;
 
-    uint64_t registerTexture(
+    // @return registered layer index
+    uint32_t registerTexture(
         const util::Ref<Texture>& texture);
 
     void updateTexture(
